+++ pallets/file-system/src/tests.rs
@@ -1,4 +1,5 @@
 use crate::{
+    self as file_system,
     mock::*,
     types::{
         AcceptedStorageRequestParameters, BatchResponses, BucketIdFor, BucketMoveRequestResponse,
@@ -1479,7 +1480,7 @@ mod request_storage {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -1552,7 +1553,7 @@ mod request_storage {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -1590,7 +1591,7 @@ mod request_storage {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -1696,7 +1697,7 @@ mod request_storage {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -1727,13 +1728,13 @@ mod request_storage {
 
                 // Assert that the next expiration block number is the storage request ttl since a single storage request was made
                 assert_eq!(
-                    FileSystem::next_available_storage_request_expiration_block(),
+                    file_system::NextAvailableStorageRequestExpirationBlock::<Test>::get(),
                     expiration_block
                 );
 
                 // Assert that the storage request expiration was appended to the list at `StorageRequestTtl`
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expiration_block),
+                    file_system::StorageRequestExpirations::<Test>::get(expiration_block),
                     vec![file_key]
                 );
 
@@ -1741,7 +1742,7 @@ mod request_storage {
 
                 // Assert that the storage request expiration was removed from the list at `StorageRequestTtl`
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expiration_block),
+                    file_system::StorageRequestExpirations::<Test>::get(expiration_block),
                     vec![]
                 );
             });
@@ -1799,7 +1800,10 @@ mod request_storage {
 
                 // Assert that the storage request expirations storage is at max capacity
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expected_expiration_block_number).len(),
+                    file_system::StorageRequestExpirations::<Test>::get(
+                        expected_expiration_block_number
+                    )
+                    .len(),
                     max_expired_items_in_block as usize
                 );
 
@@ -1808,7 +1812,9 @@ mod request_storage {
 
                 // Assert that the storage request expiration was removed from the list at `StorageRequestTtl`
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expected_expiration_block_number),
+                    file_system::StorageRequestExpirations::<Test>::get(
+                        expected_expiration_block_number
+                    ),
                     vec![]
                 );
             });
@@ -1871,7 +1877,10 @@ mod request_storage {
 
                 // Assert that the storage request expirations storage is at max capacity
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expected_expiration_block_number).len(),
+                    file_system::StorageRequestExpirations::<Test>::get(
+                        expected_expiration_block_number
+                    )
+                    .len(),
                     max_storage_request_expiry as usize
                 );
 
@@ -1882,7 +1891,10 @@ mod request_storage {
 
                 // Assert that the storage request expirations storage is at max capacity
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expected_expiration_block_number).len(),
+                    file_system::StorageRequestExpirations::<Test>::get(
+                        expected_expiration_block_number
+                    )
+                    .len(),
                     max_storage_request_expiry as usize
                 );
 
@@ -1891,13 +1903,15 @@ mod request_storage {
 
                 // Assert that the storage request expiration was removed from the list at `StorageRequestTtl`
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expected_expiration_block_number),
+                    file_system::StorageRequestExpirations::<Test>::get(
+                        expected_expiration_block_number
+                    ),
                     vec![]
                 );
 
                 // Assert that the `NextExpirationInsertionBlockNumber` storage is set to the next block number
                 assert_eq!(
-                    FileSystem::next_starting_block_to_clean_up(),
+                    file_system::NextStartingBlockToCleanUp::<Test>::get(),
                     System::block_number() + 1
                 );
             });
@@ -2009,7 +2023,7 @@ mod revoke_storage_request {
 
                 // Assert that the storage request expiration was appended to the list at `StorageRequestTtl`
                 assert_eq!(
-                    FileSystem::storage_request_expirations(expiration_block),
+                    file_system::StorageRequestExpirations::<Test>::get(expiration_block),
                     vec![file_key]
                 );
 
@@ -2067,7 +2081,7 @@ mod revoke_storage_request {
 
                 // Check StorageRequestBsps storage for confirmed BSPs
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: false,
@@ -2231,7 +2245,9 @@ mod msp_respond_storage_request {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
@@ -2342,12 +2358,16 @@ mod msp_respond_storage_request {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(first_file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(first_file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
                 assert_eq!(
-                    FileSystem::storage_requests(second_file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(second_file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
@@ -2486,13 +2506,15 @@ mod msp_respond_storage_request {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(first_file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(first_file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
 				// Assert that the storage was updated
 				assert_eq!(
-					FileSystem::storage_requests(second_file_key).unwrap().msp,
+					file_system::StorageRequests::<Test>::get(second_file_key).unwrap().msp,
 					Some((msp_id, true))
 				);
 
@@ -2650,13 +2672,17 @@ mod msp_respond_storage_request {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(first_file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(first_file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(second_file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(second_file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
@@ -2813,8 +2839,11 @@ mod msp_respond_storage_request {
                 );
 
                 // Storage request should be removed
-                assert!(FileSystem::storage_requests(file_key).is_none());
-                assert!(FileSystem::storage_request_buckets(bucket_id, file_key).is_none());
+                assert!(file_system::StorageRequests::<Test>::get(file_key).is_none());
+                assert!(
+                    file_system::BucketsWithStorageRequests::<Test>::get(bucket_id, file_key)
+                        .is_none()
+                );
             });
         }
 
@@ -4130,7 +4159,7 @@ mod bsp_volunteer {
 
                 // Assert that the RequestStorageBsps has the correct value
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: false,
@@ -4223,7 +4252,7 @@ mod bsp_volunteer {
 
                 // Assert that the RequestStorageBsps has the correct value
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: false,
@@ -4477,7 +4506,9 @@ mod bsp_confirm {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key).unwrap().msp,
+                    file_system::StorageRequests::<Test>::get(file_key)
+                        .unwrap()
+                        .msp,
                     Some((msp_id, true))
                 );
 
@@ -4715,7 +4746,7 @@ mod bsp_confirm {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -4734,7 +4765,7 @@ mod bsp_confirm {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -4851,7 +4882,7 @@ mod bsp_confirm {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -4870,7 +4901,7 @@ mod bsp_confirm {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -4954,7 +4985,7 @@ mod bsp_confirm {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: current_block,
                         owner: owner_account_id.clone(),
@@ -4973,7 +5004,7 @@ mod bsp_confirm {
 
 				// Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5132,7 +5163,7 @@ mod bsp_stop_storing {
 
                 // Assert that the RequestStorageBsps now contains the BSP under the location
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5142,7 +5173,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5251,7 +5282,7 @@ mod bsp_stop_storing {
 
                 // Assert that the RequestStorageBsps now contains the BSP under the location
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5261,7 +5292,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5388,7 +5419,7 @@ mod bsp_stop_storing {
 
                 // Assert that the RequestStorageBsps now contains the BSP under the location
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5398,7 +5429,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5439,11 +5470,11 @@ mod bsp_stop_storing {
                 ));
 
                 // Assert that the RequestStorageBsps has the correct value
-                assert!(FileSystem::storage_request_bsps(file_key, bsp_id).is_none());
+                assert!(file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id).is_none());
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5559,7 +5590,7 @@ mod bsp_stop_storing {
 
                 // Assert that the RequestStorageBsps now contains the BSP under the location
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5569,7 +5600,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5610,11 +5641,11 @@ mod bsp_stop_storing {
                 ));
 
                 // Assert that the RequestStorageBsps has the correct value
-                assert!(FileSystem::storage_request_bsps(file_key, bsp_id).is_none());
+                assert!(file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id).is_none());
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5711,7 +5742,7 @@ mod bsp_stop_storing {
 
                 // Assert that the RequestStorageBsps now contains the BSP under the location
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -5721,7 +5752,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5762,11 +5793,11 @@ mod bsp_stop_storing {
                 ));
 
                 // Assert that the RequestStorageBsps has the correct value
-                assert!(FileSystem::storage_request_bsps(file_key, bsp_id).is_none());
+                assert!(file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id).is_none());
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -5912,11 +5943,11 @@ mod bsp_stop_storing {
                 ));
 
                 // Assert that the RequestStorageBsps has the correct value
-                assert!(FileSystem::storage_request_bsps(file_key, bsp_id).is_none());
+                assert!(file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id).is_none());
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -6011,7 +6042,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage request bsps_required was incremented
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -6092,7 +6123,7 @@ mod bsp_stop_storing {
 
                 // Assert that the storage request was created with one bsps_required
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 5,
                         owner: owner_account_id.clone(),
@@ -6332,7 +6363,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was added to storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id.clone()),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id.clone()),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::try_from(
                         vec![(file_key, bucket_id)]
                     )
@@ -6360,7 +6391,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was not removed from storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::try_from(
                         vec![(file_key, bucket_id)]
                     )
@@ -6525,7 +6556,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was added to storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id.clone()),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id.clone()),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::try_from(
                         vec![(file_key, bucket_id)]
                     )
@@ -6540,7 +6571,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was added to storage
                 assert_eq!(
-                    FileSystem::file_deletion_request_expirations(expiration_block),
+                    file_system::FileDeletionRequestExpirations::<Test>::get(expiration_block),
                     vec![(
                         owner_account_id.clone(),
                         file_key
@@ -6552,13 +6583,13 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Item expiration should be removed
                 assert_eq!(
-                    FileSystem::file_deletion_request_expirations(expiration_block),
+                    file_system::FileDeletionRequestExpirations::<Test>::get(expiration_block),
                     vec![]
                 );
 
                 // Asser that the pending file deletion request was removed from storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id.clone()),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id.clone()),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::default()
                 );
 
@@ -6606,7 +6637,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was added to storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id.clone()),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id.clone()),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::try_from(
                         vec![(file_key, bucket_id)]
                     )
@@ -6646,7 +6677,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was removed from storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::default()
                 );
             });
@@ -6689,7 +6720,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was added to storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id.clone()),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id.clone()),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::try_from(
                         vec![(file_key, bucket_id)]
                     )
@@ -6731,7 +6762,7 @@ mod delete_file_and_pending_deletions_tests {
 
                 // Assert that the pending file deletion request was removed from storage
                 assert_eq!(
-                    FileSystem::pending_file_deletion_requests(owner_account_id),
+                    file_system::PendingFileDeletionRequests::<Test>::get(owner_account_id),
                     BoundedVec::<_, <Test as crate::Config>::MaxUserPendingDeletionRequests>::default()
                 );
             });
@@ -6842,7 +6873,7 @@ mod compute_threshold {
 
                 let bsp_id = Providers::get_provider_id(bsp_account_id).unwrap();
 
-                let storage_request = FileSystem::storage_requests(file_key).unwrap();
+                let storage_request = file_system::StorageRequests::<Test>::get(file_key).unwrap();
 
                 FileSystem::set_global_parameters(RuntimeOrigin::root(), None, Some(1)).unwrap();
 
@@ -7258,7 +7289,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -7277,7 +7308,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -7451,7 +7482,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -7593,7 +7624,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -7612,7 +7643,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -7810,7 +7841,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -7829,7 +7860,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
@@ -8002,7 +8033,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the storage was updated
                 assert_eq!(
-                    FileSystem::storage_requests(file_key),
+                    file_system::StorageRequests::<Test>::get(file_key),
                     Some(StorageRequestMetadata {
                         requested_at: 1,
                         owner: owner_account_id.clone(),
@@ -8021,7 +8052,7 @@ mod stop_storing_for_insolvent_user {
 
                 // Assert that the RequestStorageBsps was updated
                 assert_eq!(
-                    FileSystem::storage_request_bsps(file_key, bsp_id)
+                    file_system::StorageRequestBsps::<Test>::get(file_key, bsp_id)
                         .expect("BSP should exist in storage"),
                     StorageRequestBspsMetadata::<Test> {
                         confirmed: true,
