+++ primitives/forest-verifier/src/lib.rs
@@ -254,7 +254,7 @@ where
                 (key, TrieMutation::Add(mutation)) => {
                     trie.insert(key.as_ref(), &mutation.value)
                         .map_err(|_| "Failed to insert key into trie.")?;
-                    mutated_keys_and_values.push((key.clone(), Some(mutation.value.clone())));
+                    mutated_keys_and_values.push((*key, Some(mutation.value.clone())));
                 }
                 (key, TrieMutation::Remove(_)) => {
                     let node_value = trie
@@ -275,7 +275,7 @@ where
                     } else {
                         None
                     };
-                    mutated_keys_and_values.push((key.clone(), previous_value));
+                    mutated_keys_and_values.push((*key, previous_value));
                 }
             }
         }
