+++ xcm-simulator/Cargo.toml
@@ -26,6 +26,7 @@ scale-info = { workspace = true }
 pallet-file-system = { workspace = true }
 pallet-file-system-runtime-api = { workspace = true }
 pallet-storage-providers = { workspace = true }
+pallet-storage-providers-runtime-api = { workspace = true }
 pallet-proofs-dealer = { workspace = true }
 pallet-proofs-dealer-runtime-api = { workspace = true }
 pallet-randomness = { workspace = true }
